fix types, csv header

+++ lab1/src/BinetMultiplication.java
@@ -1,5 +1,5 @@
 class BinetMultiplication {
-    public static int operationCount = 0;
+    public static long operationCount = 0;
 
     public static double[][] multiply(double[][] A, double[][] B) {
         operationCount = 0;
